Remove Google Translate support
it's not production proof
