+++ docs/_layouts/default.html
@@ -22,30 +22,26 @@
               </div>
               <div class="dropdown-menu" id="bento-menu" role="menu">
                 <div class="dropdown-content">
-                  <a class="dropdown-item" href="https://translatiumapp.com">
-                    <img src="/images/logo.png" alt="Translatium">
-                    <p>Translate Any Languages like a Pro.</p>
-                  </a>
-                  <hr class="dropdown-divider">
                   <a class="dropdown-item" href="https://getwebcatalog.com">
-                    <img src="/images/webcatalog-logo.png" alt="WebCatalog">
+                    <img src="/images/webcatalog-logo.svg" alt="WebCatalog">
                     <p>Run Web Apps like Real Apps.</p>
                   </a>
                   <hr class="dropdown-divider">
                   <a class="dropdown-item" href="https://singleboxapp.com">
-                    <img src="/images/singlebox-logo.png" alt="Singlebox">
+                    <img src="/images/singlebox-logo.svg" alt="Singlebox">
                     <p>All Your Apps in One Single Window.</p>
                   </a>
                   <hr class="dropdown-divider">
-                  <a class="dropdown-item" href="https://quanglam2807.com">
-                    <img src="/images/quanglam-logo.png" alt="Quang Lam">
+                  <a class="dropdown-item" href="https://translatiumapp.com">
+                    <img src="/images/translatium-logo.svg" alt="Translatium">
+                    <p>Translate Any Languages like a Pro.</p>
                   </a>
                 </div>
               </div>
             </div>
           </div>
           <a class="navbar-item" href="/">
-            <img src="{{ '/images/logo.png' | relative_url }}" alt="Translatium">
+            <img src="{{ '/images/translatium-logo.svg' | relative_url }}" alt="Translatium">
           </a>
           <span class="navbar-burger burger" data-target="navbarMenu" aria-label="menu" aria-expanded="false">
             <span></span>
