+++ docs/_layouts/download.html
@@ -4,7 +4,7 @@ layout: default
 <div class="hero-body">
   <div class="container">
     <div class="has-text-centered">
-      <h1 class="title is-2">
+      <h1 class="title is-3">
         Translate Any Languages like a Pro
       </h1>
       <h2 class="subtitle">
@@ -109,7 +109,7 @@ layout: default
         Features
       </h2>
       <nav class="columns is-multiline">
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-language"></i>
@@ -117,7 +117,7 @@ layout: default
             <p><strong><a href="/languages">90+ Languages</a></strong><br>90+ languages & dialects.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-globe"></i>
@@ -125,7 +125,7 @@ layout: default
             <p><strong>Auto Detect Language</strong><br>Identify languages automatically.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-headphones"></i>
@@ -133,7 +133,7 @@ layout: default
             <p><strong>Text-To-Speech</strong><br/>Listen to your translations.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-camera"></i>
@@ -141,7 +141,7 @@ layout: default
             <p><strong>Camera Translation</strong><br>Take a picture to translate.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-book"></i>
@@ -149,7 +149,7 @@ layout: default
             <p><strong>Dictionary</strong><br>Synonyms, examples and more.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-history"></i>
@@ -157,7 +157,7 @@ layout: default
             <p><strong>History</strong><br>Access to previous translations.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-star"></i>
@@ -165,31 +165,15 @@ layout: default
             <p><strong>Phrasebook</strong><br>Save your favorite translations.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
-          <div class="has-text-centered">
-            <span class="icon is-large">
-              <i class="fa fa-2x fa-expand"></i>
-            </span>
-            <p><strong>Fullscreen</strong><br>View translations in fullscreen.</p>
-          </div>
-        </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-paint-brush"></i>
             </span>
-            <p><strong>Personalization</strong><br>With beautiful themes and colors.</p>
+            <p><strong>Personalization</strong><br>Beautiful dark theme and more.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
-          <div class="has-text-centered">
-            <span class="icon is-large">
-              <i class="fa fa-2x fa-columns"></i>
-            </span>
-            <p><strong>Snap View</strong><br>Run side by side with other apps.</p>
-          </div>
-        </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-bars"></i>
@@ -197,7 +181,7 @@ layout: default
             <p><strong>Stay in Menu Bar</strong><br>Access quickly from anywhere.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-keyboard"></i>
@@ -205,7 +189,7 @@ layout: default
             <p><strong>Keyboard Shortcuts</strong><br>Be productive!</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-clipboard"></i>
@@ -213,7 +197,7 @@ layout: default
             <p><strong>Clipboard</strong><br>Easily copy and share translations.</p>
           </div>
         </div>
-        <div class="column is-half-tablet is-half-desktop">
+        <div class="column is-half-tablet is-one-third-desktop">
           <div class="has-text-centered">
             <span class="icon is-large">
               <i class="fa fa-2x fa-exchange-alt"></i>
