+++ docs/_sass/bulma.sass
@@ -22,9 +22,6 @@ body
 main
   flex: 1
 
-.container
-  max-width: 800px !important
-
 .footer
   padding: 1.5em
 
@@ -39,12 +36,6 @@ main
 .is-text-link
   &:hover
     text-decoration: underline !important
-
-.app-icon
-  max-width: 96px
-
-.app-icon-small
-  max-width: 56px
 
 small.price-end
   font-size: 0.7em
