+++ public/libs/locales/en.json
@@ -90,7 +90,6 @@
 	"close": "Close",
 	"general": "General",
 	"advanced": "Advanced",
-	"preferredTranslationService": "Preferred translation provider",
 	"alwaysOnTop": "Keep window always on top",
 	"alwaysOnTopDesc": "Translatium will not hide when you click outside.",
 	"translateClipboardOnShortcut": "Translate clipboard on shortcut",
@@ -121,24 +120,10 @@
 	"examples": "Examples",
 	"seeAlso": "See also",
 	"translations": "Translations",
+	"auto": "Auto detect",
 	"af": "Afrikaans",
 	"am": "Amharic",
 	"ar": "Arabic",
-	"ar-AE": "Arabic (United Arab Emirates)",
-	"ar-BH": "Arabic (Bahrain)",
-	"ar-DZ": "Arabic (Algeria)",
-	"ar-EG": "Arabic (Egypt)",
-	"ar-IL": "Arabic (Israel)",
-	"ar-JO": "Arabic (Jordan)",
-	"ar-KW": "Arabic (Kuwait)",
-	"ar-LB": "Arabic (Lebanon)",
-	"ar-MA": "Arabic (Morocco)",
-	"ar-OM": "Arabic (Oman)",
-	"ar-PS": "Arabic (Palestine)",
-	"ar-QA": "Arabic (Qatar)",
-	"ar-SA": "Arabic (Saudi Arabia)",
-	"ar-TN": "Arabic (Tunisia)",
-	"auto": "Auto detect",
 	"az": "Azerbaijani",
 	"ba": "Bashkir",
 	"be": "Belarusian",
@@ -147,77 +132,39 @@
 	"bs": "Bosnian",
 	"ca": "Catalan",
 	"ceb": "Cebuano",
-	"co": "Corsican",
 	"cs": "Czech",
 	"cy": "Welsh",
 	"da": "Danish",
 	"de": "German",
 	"el": "Greek",
 	"en": "English",
-	"en-AU": "English (Australia)",
-	"en-CA": "English (Canada)",
-	"en-IE": "English (Ireland)",
-	"en-IN": "English (India)",
-	"en-NZ": "English (New Zealand)",
-	"en-PH": "English (Phillipines)",
-	"en-UK": "English (United Kingdom)",
-	"en-US": "English (United States)",
-	"en-ZA": "English (South Africa)",
 	"eo": "Esperanto",
 	"es": "Spanish",
-	"es-AR": "Spanish (Argentina)",
-	"es-BO": "Spanish (Bolivia)",
-	"es-CL": "Spanish (Chile)",
-	"es-CO": "Spanish (Colombia)",
-	"es-CR": "Spanish (Costa Rica)",
-	"es-DO": "Spanish (Dominican Republic)",
-	"es-EC": "Spanish (Ecuador)",
-	"es-ES": "Spanish (Spain)",
-	"es-GT": "Spanish (Guatemala)",
-	"es-HN": "Spanish (Honduras)",
-	"es-MX": "Spanish (Mexico)",
-	"es-NI": "Spanish (Nicaragua)",
-	"es-PA": "Spanish (Panama)",
-	"es-PE": "Spanish (Peru)",
-	"es-PR": "Spanish (Puerto Rico)",
-	"es-PY": "Spanish (Paraguay)",
-	"es-SV": "Spanish (El Salvador)",
-	"es-US": "Spanish (United States)",
-	"es-UY": "Spanish (Uruguay)",
-	"es-VE": "Spanish (Venezuela)",
 	"et": "Estonian",
 	"eu": "Basque",
 	"fa": "Persian",
 	"fi": "Finnish",
 	"fr": "French",
-	"fy": "Frisian",
 	"ga": "Irish",
 	"gd": "Scottish Gaelic",
 	"gl": "Galician",
 	"gu": "Gujarati",
-	"ha": "Hausa",
-	"haw": "Hawaiian",
 	"he": "Hebrew",
 	"hi": "Hindi",
-	"hmn": "Hmong",
 	"hr": "Croatian",
-	"ht": "Haitian Creole",
+	"ht": "Haitian",
 	"hu": "Hungarian",
 	"hy": "Armenian",
 	"id": "Indonesian",
-	"ig": "Igbo",
 	"is": "Icelandic",
 	"it": "Italian",
-	"iw": "Hebrew",
 	"ja": "Japanese",
 	"jv": "Javanese",
-	"jw": "Javanese",
 	"ka": "Georgian",
 	"kk": "Kazakh",
 	"km": "Khmer",
 	"kn": "Kannada",
 	"ko": "Korean",
-	"ku": "Kurdish (Kurmanji)",
 	"ky": "Kyrgyz",
 	"la": "Latin",
 	"lb": "Luxembourgish",
@@ -234,30 +181,21 @@
 	"mrj": "Hill Mari",
 	"ms": "Malay",
 	"mt": "Maltese",
-	"my": "Myanmar (Burmese)",
+	"my": "Burmese",
 	"ne": "Nepali",
 	"nl": "Dutch",
 	"no": "Norwegian",
-	"ny": "Chichewa",
 	"pa": "Punjabi",
 	"pap": "Papiamento",
 	"pl": "Polish",
-	"ps": "Pashto",
 	"pt": "Portuguese",
-	"pt-BR": "Portuguese (Brazil)",
-	"pt-PT": "Portuguese (Portugal)",
 	"ro": "Romanian",
 	"ru": "Russian",
-	"sd": "Sindhi",
-	"si": "Sinhala",
+	"si": "Sinhalese",
 	"sk": "Slovak",
 	"sl": "Slovenian",
-	"sm": "Samoan",
-	"sn": "Shona",
-	"so": "Somali",
 	"sq": "Albanian",
 	"sr": "Serbian",
-	"st": "Sesotho",
 	"su": "Sundanese",
 	"sv": "Swedish",
 	"sw": "Swahili",
@@ -265,7 +203,7 @@
 	"te": "Telugu",
 	"tg": "Tajik",
 	"th": "Thai",
-	"tl": "Filipino",
+	"tl": "Tagalog",
 	"tr": "Turkish",
 	"tt": "Tatar",
 	"udm": "Udmurt",
@@ -275,10 +213,7 @@
 	"vi": "Vietnamese",
 	"xh": "Xhosa",
 	"yi": "Yiddish",
-	"yo": "Yoruba",
 	"zh": "Chinese",
-	"zh-CN": "Chinese (Simplified, China)",
-	"zh-HK": "Chinese (Traditional, Hong Kong)",
-	"zh-TW": "Chinese (Traditional, Taiwan)",
-	"zu": "Zulu"
+	"zh-CN": "Chinese (Simplified)",
+	"zh-TW": "Chinese (Traditional)"
 }
