+++ public/libs/locales/vi.json
@@ -90,7 +90,6 @@
 	"close": "Đóng",
 	"general": "Cài đặt chung",
 	"advanced": "Cài đặt nâng cao",
-	"preferredTranslationService": "nhà cung cấp dịch ưu tiên",
 	"alwaysOnTop": "Giữ cửa sổ luôn ở trên cùng",
 	"alwaysOnTopDesc": "Cửa sổ sẽ không ẩn khi bạn rời chuột đi.",
 	"translateClipboardOnShortcut": "Dịch clipboard khi dùng phím tắt",
@@ -121,24 +120,10 @@
 	"examples": "Ví dụ",
 	"seeAlso": "Xem thêm",
 	"translations": "Dịch",
+	"auto": "Tự động nhận diện",
 	"af": "Tiếng Afrikaans",
 	"am": "Tiếng Amhara",
 	"ar": "Tiếng Ả Rập",
-	"ar-AE": "Tiếng Ả Rập (Các Tiểu vương quốc Ả Rập Thống nhất)",
-	"ar-BH": "Tiếng Ả Rập (Bahrain)",
-	"ar-DZ": "Tiếng Ả Rập (Algeria)",
-	"ar-EG": "Tiếng Ả Rập (Ai Cập)",
-	"ar-IL": "Tiếng Ả Rập (Israel)",
-	"ar-JO": "Tiếng Ả Rập (Jordan)",
-	"ar-KW": "Tiếng Ả Rập (Kuwait)",
-	"ar-LB": "Tiếng Ả Rập (Lebanon)",
-	"ar-MA": "Tiếng Ả Rập (Morocco)",
-	"ar-OM": "Tiếng Ả Rập (Oman)",
-	"ar-PS": "Tiếng Ả Rập (Palestine)",
-	"ar-QA": "Tiếng Ả Rập (Qatar)",
-	"ar-SA": "Tiếng Ả Rập (Ả Rập Saudi)",
-	"ar-TN": "Tiếng Ả Rập (Tunisia)",
-	"auto": "Tự động nhận diện",
 	"az": "Tiếng Azerbaijan",
 	"ba": "Tiếng Bashkir",
 	"be": "Tiếng Belarus",
@@ -147,77 +132,39 @@
 	"bs": "Tiếng Bosnia",
 	"ca": "Tiếng Catalan",
 	"ceb": "Tiếng Cebuano",
-	"co": "Tiếng Corsica",
 	"cs": "Tiếng Séc",
 	"cy": "Tiếng Wales",
 	"da": "Tiếng Đan Mạch",
 	"de": "Tiếng Đức",
 	"el": "Tiếng Hy Lạp",
 	"en": "Tiếng Anh",
-	"en-AU": "Tiếng Anh (Australia)",
-	"en-CA": "Tiếng Anh (Canada)",
-	"en-IE": "Tiếng Anh (Ireland)",
-	"en-IN": "Tiếng Anh (Ấn Độ)",
-	"en-NZ": "Tiếng Anh (New Zealand)",
-	"en-PH": "Tiếng Anh (Philippines)",
-	"en-UK": "Tiếng Anh (Anh Quốc)",
-	"en-US": "Tiếng Anh (Hoa Kỳ)",
-	"en-ZA": "Tiếng Anh (Nam Phi)",
 	"eo": "Tiếng Esperanto",
 	"es": "Tiếng Tây Ban Nha",
-	"es-AR": "Tiếng Tây Ban Nha (Argentina)",
-	"es-BO": "Tiếng Tây Ban Nha (Bolivia)",
-	"es-CL": "Tiếng Tây Ban Nha (Chile)",
-	"es-CO": "Tiếng Tây Ban Nha (Colombia)",
-	"es-CR": "Tiếng Tây Ban Nha (Costa Rica)",
-	"es-DO": "Tiếng Tây Ban Nha (Cộng hòa Dominica)",
-	"es-EC": "Tiếng Tây Ban Nha (Ecuador)",
-	"es-ES": "Tiếng Tây Ban Nha (Tây Ban Nha)",
-	"es-GT": "Tiếng Tây Ban Nha (Guatemala)",
-	"es-HN": "Tiếng Tây Ban Nha (Honduras)",
-	"es-MX": "Tiếng Tây Ban Nha (Mexico)",
-	"es-NI": "Tiếng Tây Ban Nha (Nicaragua)",
-	"es-PA": "Tiếng Tây Ban Nha (Panama)",
-	"es-PE": "Tiếng Tây Ban Nha (Peru)",
-	"es-PR": "Tiếng Tây Ban Nha (Puerto Rico)",
-	"es-PY": "Tiếng Tây Ban Nha (Paraguay)",
-	"es-SV": "Tiếng Tây Ban Nha (El Salvador)",
-	"es-US": "Tiếng Tây Ban Nha (Hoa Kỳ)",
-	"es-UY": "Tiếng Tây Ban Nha (Uruguay)",
-	"es-VE": "Tiếng Tây Ban Nha (Venezuela)",
 	"et": "Tiếng Estonia",
 	"eu": "Tiếng Basque",
 	"fa": "Tiếng Persia",
 	"fi": "Tiếng Phần Lan",
 	"fr": "Tiếng Pháp",
-	"fy": "Tiếng Frisian",
 	"ga": "Tiếng Ireland",
 	"gd": "Tiếng Gael Scotland",
 	"gl": "Tiếng Galicia",
 	"gu": "Tiếng Gujarat",
-	"ha": "Tiếng Hausa",
-	"haw": "Tiếng Hawaii",
 	"he": "Tiếng Hebrew",
 	"hi": "Tiếng Hindi",
-	"hmn": "Tiếng Hmong",
 	"hr": "Tiếng Croatia",
 	"ht": "Tiếng Haitia",
 	"hu": "Tiếng Hungaria",
 	"hy": "Tiếng Armenia",
 	"id": "Tiếng Indonesia",
-	"ig": "Tiếng Igbo",
 	"is": "Tiếng Iceland",
 	"it": "Tiếng Ý",
-	"iw": "Tiếng Hebrew",
 	"ja": "Tiếng Nhật",
 	"jv": "Tiếng Javan",
-	"jw": "Tiếng Java",
 	"ka": "Tiếng Georgia",
 	"kk": "Tiếng Kazakh",
 	"km": "Tiếng Khmer",
 	"kn": "Tiếng Kannada",
 	"ko": "Tiếng Hàn",
-	"ku": "Tiếng Kurd (Kurmanji)",
 	"ky": "Tiếng Kyrgyz",
 	"la": "Tiếng Latin",
 	"lb": "Tiếng Luxembourg",
@@ -238,26 +185,17 @@
 	"ne": "Tiếng Nepal",
 	"nl": "Tiếng Hà Lan",
 	"no": "Tiếng Na Uy",
-	"ny": "Tiếng Chichewa",
 	"pa": "Tiếng Punjab",
 	"pap": "Tiếng Papiamento",
 	"pl": "Tiếng Ba Lan",
-	"ps": "Tiếng Pashto",
 	"pt": "Tiếng Bồ Đào Nha",
-	"pt-BR": "Tiếng Bồ Đào Nha (Brazil)",
-	"pt-PT": "Tiếng Bồ Đào Nha (Bồ Đào Nha)",
 	"ro": "Tiếng Romania",
 	"ru": "Tiếng Nga",
-	"sd": "Tiếng Sindhi",
 	"si": "Tiếng Sinhala",
 	"sk": "Tiếng Slovak",
 	"sl": "Tiếng Slovenia",
-	"sm": "Tiếng Samoa",
-	"sn": "Tiếng Shona",
-	"so": "Tiếng Somali",
 	"sq": "Tiếng Albania",
 	"sr": "Tiếng Serbia",
-	"st": "Tiếng Sesotho",
 	"su": "Tiếng Sundan",
 	"sv": "Tiếng Thuỵ Điển",
 	"sw": "Tiếng Swahili",
@@ -275,10 +213,7 @@
 	"vi": "Tiếng Việt",
 	"xh": "Tiếng Xhosa",
 	"yi": "Tiếng Yiddish",
-	"yo": "Tiếng Yoruba",
 	"zh": "Tiếng Trung",
-	"zh-CN": "Tiếng Trung (Giản thể, Trung Quốc)",
-	"zh-HK": "Tiếng Trung (Giản thể, Hồng Kông)",
-	"zh-TW": "Tiếng Trung (Phồn thể, Đài Loan)",
-	"zu": "Tiếng Zulu"
+	"zh-CN": "Tiếng Trung (Giản thể)",
+	"zh-TW": "Tiếng Trung (Phồn thể)"
 }
