+++ public/libs/locales/zh-CN.json
@@ -4,48 +4,48 @@
 	"redo": "重做",
 	"cut": "切",
 	"copy": "拷贝",
-	"paste": "糊",
+	"paste": "贴",
 	"delete": "删除",
-	"selectAll": "全选",
-	"view": "视图",
+	"selectAll": "所有的选择",
+	"view": "看",
 	"toggleFullscreen": "切换全屏",
-	"toggleDevTools": "切换开发者工具",
+	"toggleDevTools": "切换开发工具",
 	"window": "窗口",
 	"minimize": "最小化",
 	"zoom": "放大",
-	"bringAllToFront": "前置全部",
-	"help": "救命",
-	"learnMore": "了解更多关于Translatium ...",
-	"reportAnIssue": "报告问题...",
+	"bringAllToFront": "把所有以前",
+	"help": "帮助",
+	"learnMore": "了解更多关于Translatium...",
+	"reportAnIssue": "报告的一个问题...",
 	"file": "文件",
 	"about": "关于",
 	"checkForUpdates": "检查更新...",
 	"hide": "隐藏Translatium",
-	"hideOthers": "隐藏其他",
-	"unhide": "取消隐藏",
+	"hideOthers": "隐藏其他人",
+	"unhide": "隐藏",
 	"quit": "退出",
 	"home": "主页",
 	"phrasebook": "收藏夹",
 	"preferences": "设置",
-	"preferencesMenuItem": "优先...",
-	"ok": "好",
-	"updateAvailable": "有可用的更新",
-	"updateAvailableDesc": "有可用的更新。它正在下载。我们将让你知道什么时候准备好",
+	"preferencesMenuItem": "喜好...",
+	"ok": "\"确定\"",
+	"updateAvailable": "更新可用",
+	"updateAvailableDesc": "有可用的更新。 它正在下载。 我们将让你知道什么时候准备好",
 	"noUpdates": "没有更新",
-	"noUpdatesDesc": "目前没有可用的更新。",
+	"noUpdatesDesc": "目前没有更新。",
 	"failedToCheckForUpdates": "无法检查更新",
-	"failedToCheckForUpdatesDesc": "无法检查更新。请检查您的Internet连接。",
-	"restart": "重新开始",
-	"later": "后来",
-	"applicationUpdate": "应用程序更新",
-	"applicationUpdateDesc": "新版本$VERSION已下载。重新启动以应用更新的应用程序。",
-	"resetNow": "立即重置",
-	"resetDesc": "你确定吗？所有优惠将恢复为原始默认设置。此操作无法撤消。",
-	"requireRestartDesc": "您需要手动退出并重新启动应用程序，此更改生效。",
+	"failedToCheckForUpdatesDesc": "无法检查更新。 请检查你的连接互联网。",
+	"restart": "重新启动",
+	"later": "后",
+	"applicationUpdate": "更新应用程序",
+	"applicationUpdateDesc": "一个新的版本$版已被下载。 重新启动的应用程序申请的更新。",
+	"resetNow": "重现在",
+	"resetDesc": "你确定吗？ 所有的喜好会被恢复到其原有违约。 这种行动不能被撤销。",
+	"requireRestartDesc": "你需要退出然后手工重新启动该程序这一更改的影响。",
 	"quitNow": "现在退出",
 	"isRequired": "$FIELDNAME是必需的。",
-	"isNotValid": "$FIELDNAME无效。",
-	"website": "Translatium 官网",
+	"isNotValid": "$FIELDNAME不是有效的。",
+	"website": "官网",
 	"support": "支持",
 	"typeSomethingHere": "请输入翻译内容...",
 	"clear": "清空",
@@ -90,13 +90,12 @@
 	"close": "关闭",
 	"general": "一般",
 	"advanced": "高级",
-	"preferredTranslationService": "首选翻译服务提供商",
 	"alwaysOnTop": "保持窗口始终在顶部",
 	"alwaysOnTopDesc": "窗口不会隐藏在模糊上。",
 	"translateClipboardOnShortcut": "快捷翻译",
 	"translateClipboardOnShortcutDesc": "当您使用快捷方式打开 Translatium 时，它将自动翻译您剪贴板的内容。",
-	"openKeyboardShortcut": "快速定义键盘快捷键打开Translatium",
-	"typeNewKeyboardCombinator": "键入新的键盘组合子。",
+	"openKeyboardShortcut": "定义一个快捷键打开Translatium快",
+	"typeNewKeyboardCombinator": "类型的新的键盘组合.",
 	"searchLanguages": "搜索语言",
 	"noLanguageFound": "没有找到对应语言",
 	"searchResults": "搜索结果",
@@ -105,7 +104,7 @@
 	"remove": "移除",
 	"andSaveToHistory": "翻译并保存到历史记录",
 	"save": "保存",
-	"removeShortcut": "删除快捷方式",
+	"removeShortcut": "除快捷方式",
 	"cancel": "取消",
 	"pasteFromClipboard": "从剪贴板粘贴",
 	"popclipExtension": "PopClip Extension",
@@ -117,28 +116,14 @@
 	"visitStore": "访问商店。。。",
 	"register": "注册",
 	"definitions": "定义",
-	"synonyms": "近义词",
-	"examples": "范例",
-	"seeAlso": "参看",
+	"synonyms": "同义词",
+	"examples": "例",
+	"seeAlso": "也参看",
 	"translations": "翻译",
+	"auto": "自动检测",
 	"af": "南非荷兰语",
 	"am": "阿姆哈拉语",
 	"ar": "阿拉伯语",
-	"ar-AE": "阿拉伯语 (阿拉伯联合酋长国)",
-	"ar-BH": "阿拉伯语 (巴林)",
-	"ar-DZ": "阿拉伯语 (阿尔及利亚)",
-	"ar-EG": "阿拉伯语 (埃及)",
-	"ar-IL": "阿拉伯语 (以色列)",
-	"ar-JO": "阿拉伯语 (约旦)",
-	"ar-KW": "阿拉伯语 (科威特)",
-	"ar-LB": "阿拉伯语 (黎巴嫩)",
-	"ar-MA": "阿拉伯语 (摩洛哥)",
-	"ar-OM": "阿拉伯语 (阿曼)",
-	"ar-PS": "阿拉伯语 (巴勒斯坦)",
-	"ar-QA": "阿拉伯语 (卡塔尔)",
-	"ar-SA": "阿拉伯语 (沙特阿拉伯)",
-	"ar-TN": "阿拉伯语 (突尼斯)",
-	"auto": "自动检测",
 	"az": "阿塞拜疆语",
 	"ba": "巴什基尔",
 	"be": "白俄罗斯语",
@@ -147,77 +132,39 @@
 	"bs": "波斯尼亚语",
 	"ca": "加泰罗尼亚语",
 	"ceb": "宿雾语",
-	"co": "科西嘉语",
 	"cs": "捷克语",
 	"cy": "威尔士语",
 	"da": "丹麦语",
 	"de": "德语",
 	"el": "希腊语",
 	"en": "英语",
-	"en-AU": "英语 (澳大利亚)",
-	"en-CA": "英语 (加拿大)",
-	"en-IE": "英语 (爱尔兰)",
-	"en-IN": "英语 (印度)",
-	"en-NZ": "英语 (新西兰)",
-	"en-PH": "英语 (菲律宾)",
-	"en-UK": "英语 (英国)",
-	"en-US": "英语 (美国)",
-	"en-ZA": "英语 (南非)",
 	"eo": "世界语",
 	"es": "西班牙语",
-	"es-AR": "西班牙语 (阿根廷)",
-	"es-BO": "西班牙语 (玻利维亚)",
-	"es-CL": "西班牙语 (智利)",
-	"es-CO": "西班牙语 (哥伦比亚)",
-	"es-CR": "西班牙语 (哥斯达黎加)",
-	"es-DO": "西班牙语 (多米尼加共和国)",
-	"es-EC": "西班牙语 (厄瓜多尔)",
-	"es-ES": "西班牙语 (西班牙)",
-	"es-GT": "西班牙语 (危地马拉)",
-	"es-HN": "西班牙语 (洪都拉斯)",
-	"es-MX": "西班牙语 (墨西哥)",
-	"es-NI": "西班牙语 (尼加拉瓜)",
-	"es-PA": "西班牙语 (巴拿马)",
-	"es-PE": "西班牙语 (秘鲁)",
-	"es-PR": "西班牙语 (波多黎各)",
-	"es-PY": "西班牙语 (巴拉圭)",
-	"es-SV": "西班牙语 (萨尔瓦多)",
-	"es-US": "西班牙语 (美国)",
-	"es-UY": "西班牙语 (乌拉圭)",
-	"es-VE": "西班牙语 (委内瑞拉)",
 	"et": "爱沙尼亚语",
 	"eu": "巴斯克语",
 	"fa": "波斯语",
 	"fi": "芬兰语",
 	"fr": "法语",
-	"fy": "弗里西语",
 	"ga": "爱尔兰语",
 	"gd": "苏格兰盖尔语",
 	"gl": "加利西亚语",
 	"gu": "古吉拉特语",
-	"ha": "豪萨语",
-	"haw": "夏威夷语",
 	"he": "希伯来语",
 	"hi": "印地语",
-	"hmn": "苗语",
 	"hr": "克罗地亚语",
 	"ht": "海地克里奥尔语",
 	"hu": "匈牙利语",
 	"hy": "亚美尼亚语",
 	"id": "印度尼西亚语",
-	"ig": "伊博语",
 	"is": "冰岛语",
 	"it": "意大利语",
-	"iw": "希伯来语",
 	"ja": "日语",
 	"jv": "爪哇语",
-	"jw": "爪哇语",
 	"ka": "格鲁吉亚语",
 	"kk": "哈萨克语",
 	"km": "高棉语",
 	"kn": "坎那达语",
 	"ko": "朝鲜语",
-	"ku": "库尔德语",
 	"ky": "柯尔克孜语/吉尔吉斯语",
 	"la": "拉丁语",
 	"lb": "卢森堡语",
@@ -238,26 +185,17 @@
 	"ne": "尼泊尔语",
 	"nl": "荷兰语",
 	"no": "挪威语",
-	"ny": "齐佩瓦语",
 	"pa": "旁遮普语",
 	"pap": "帕皮阿门托",
 	"pl": "波兰语",
-	"ps": "普什图语",
 	"pt": "葡萄牙语",
-	"pt-BR": "葡萄牙语 (巴西)",
-	"pt-PT": "葡萄牙语 (葡萄牙)",
 	"ro": "罗马尼亚语",
 	"ru": "俄语",
-	"sd": "信德语",
 	"si": "僧伽罗语",
 	"sk": "斯洛伐克语",
 	"sl": "斯洛维尼亚语",
-	"sm": "萨摩亚语",
-	"sn": "绍纳语",
-	"so": "索马里语",
 	"sq": "阿尔巴尼亚语",
 	"sr": "塞尔维亚语",
-	"st": "塞索托语",
 	"su": "巽他语",
 	"sv": "瑞典语",
 	"sw": "斯瓦希里语",
@@ -275,10 +213,7 @@
 	"vi": "越南语",
 	"xh": "科萨语",
 	"yi": "意第绪语",
-	"yo": "约鲁巴语",
 	"zh": "中文",
-	"zh-CN": "中国（简体，中国）",
-	"zh-HK": "中国（繁体，香港）",
-	"zh-TW": "中国（繁体，台灣）",
-	"zu": "祖鲁语"
+	"zh-CN": "中文(简体)",
+	"zh-TW": "中文(繁体)"
 }
