+++ public/libs/locales/zh-TW.json
@@ -4,47 +4,47 @@
 	"redo": "重做",
 	"cut": "切",
 	"copy": "拷貝",
-	"paste": "糊",
+	"paste": "貼",
 	"delete": "刪除",
-	"selectAll": "全選",
-	"view": "視圖",
+	"selectAll": "所有的選擇",
+	"view": "看",
 	"toggleFullscreen": "切換全屏",
-	"toggleDevTools": "切換開發者工具",
+	"toggleDevTools": "切換開發工具",
 	"window": "窗口",
 	"minimize": "最小化",
 	"zoom": "放大",
-	"bringAllToFront": "前置全部",
-	"help": "救命",
-	"learnMore": "了解更多關於Translatium ...",
-	"reportAnIssue": "報告問題...",
+	"bringAllToFront": "把所有以前",
+	"help": "幫助",
+	"learnMore": "瞭解更多關於Translatium...",
+	"reportAnIssue": "報告的一個問題...",
 	"file": "文件",
 	"about": "關於",
 	"checkForUpdates": "檢查更新...",
 	"hide": "隱藏Translatium",
-	"hideOthers": "隱藏其他",
-	"unhide": "取消隱藏",
+	"hideOthers": "隱藏其他人",
+	"unhide": "隱藏",
 	"quit": "退出",
 	"home": "主頁",
 	"phrasebook": "收藏夾",
 	"preferences": "設置",
-	"preferencesMenuItem": "優先...",
-	"ok": "好",
-	"updateAvailable": "有可用的更新",
-	"updateAvailableDesc": "有可用的更新。它正在下載。我們將讓你知道什麼時候準備好",
+	"preferencesMenuItem": "喜好...",
+	"ok": "\"確定\"",
+	"updateAvailable": "更新可用",
+	"updateAvailableDesc": "有可用的更新。 它正在下載。 我們將讓你知道什麼時候准備好",
 	"noUpdates": "沒有更新",
-	"noUpdatesDesc": "目前沒有可用的更新。",
+	"noUpdatesDesc": "目前沒有更新。",
 	"failedToCheckForUpdates": "無法檢查更新",
-	"failedToCheckForUpdatesDesc": "無法檢查更新。請檢查您的Internet連接。",
-	"restart": "重新開始",
-	"later": "後來",
-	"applicationUpdate": "應用程序更新",
-	"applicationUpdateDesc": "新版本$VERSION已下載。重新啟動以應用更新的應用程序。",
-	"resetNow": "立即重置",
-	"resetDesc": "你確定嗎？所有優惠將恢復為原始默認設置。此操作無法撤消。",
-	"requireRestartDesc": "您需要手動退出並重新啟動應用程序，此更改生效。",
+	"failedToCheckForUpdatesDesc": "無法檢查更新。 請檢查你的連接互聯網。",
+	"restart": "重新啟動",
+	"later": "後",
+	"applicationUpdate": "更新應用程序",
+	"applicationUpdateDesc": "一個新的版本$版已被下載。 重新啟動的應用程序申請的更新。",
+	"resetNow": "重現在",
+	"resetDesc": "你確定嗎？ 所有的喜好會被恢復到其原有違約。 這種行動不能被撤銷。",
+	"requireRestartDesc": "你需要退出然後手工重新啟動該程序這一更改的影響。",
 	"quitNow": "現在退出",
 	"isRequired": "$FIELDNAME是必需的。",
-	"isNotValid": "$FIELDNAME無效。",
+	"isNotValid": "$FIELDNAME不是有效的。",
 	"website": "官網",
 	"support": "支持",
 	"typeSomethingHere": "請輸入翻譯內容...",
@@ -90,13 +90,12 @@
 	"close": "關閉",
 	"general": "一般",
 	"advanced": "高級",
-	"preferredTranslationService": "首選翻譯服務提供商",
 	"alwaysOnTop": "保持窗口始終在頂部",
 	"alwaysOnTopDesc": "窗口不會隱藏在模糊上。",
 	"translateClipboardOnShortcut": "快捷翻譯",
 	"translateClipboardOnShortcutDesc": "當您使用快捷方式打開 Translatium 時，它將自動翻譯您剪貼板的內容。",
-	"openKeyboardShortcut": "快速定義鍵盤快捷鍵打開Translatium",
-	"typeNewKeyboardCombinator": "鍵入新的鍵盤組合子。",
+	"openKeyboardShortcut": "定義一個快捷鍵打開Translatium快",
+	"typeNewKeyboardCombinator": "類型的新的鍵盤組合.",
 	"searchLanguages": "搜索語言",
 	"noLanguageFound": "沒有找到對應語言",
 	"searchResults": "搜索結果",
@@ -105,7 +104,7 @@
 	"remove": "移除",
 	"andSaveToHistory": "翻譯並保存到歷史記錄",
 	"save": "保存",
-	"removeShortcut": "刪除快捷方式",
+	"removeShortcut": "除快捷方式",
 	"cancel": "取消",
 	"pasteFromClipboard": "從剪貼板粘貼",
 	"popclipExtension": "PopClip Extension",
@@ -118,27 +117,13 @@
 	"register": "註冊",
 	"definitions": "定義",
 	"synonyms": "同義詞",
-	"examples": "例子",
-	"seeAlso": "也可以看看",
+	"examples": "例",
+	"seeAlso": "也參看",
 	"translations": "翻譯",
+	"auto": "自動檢測",
 	"af": "南非荷蘭語",
 	"am": "阿姆哈拉語",
 	"ar": "阿拉伯語",
-	"ar-AE": "阿拉伯語（阿拉伯聯合酋長國）",
-	"ar-BH": "阿拉伯語（巴林）",
-	"ar-DZ": "阿拉伯語（阿爾及利亞）",
-	"ar-EG": "阿拉伯語（埃及）",
-	"ar-IL": "阿拉伯語（以色列）",
-	"ar-JO": "阿拉伯語（約旦）",
-	"ar-KW": "阿拉伯語（科威特）",
-	"ar-LB": "阿拉伯語（黎巴嫩）",
-	"ar-MA": "阿拉伯語（摩洛哥）",
-	"ar-OM": "阿拉伯語（阿曼）",
-	"ar-PS": "阿拉伯語（巴勒斯坦）",
-	"ar-QA": "阿拉伯語（卡塔爾）",
-	"ar-SA": "阿拉伯語（沙特阿拉伯）",
-	"ar-TN": "阿拉伯語（突尼斯）",
-	"auto": "自動檢測",
 	"az": "阿塞拜疆語",
 	"ba": "巴什基爾",
 	"be": "白俄羅斯語",
@@ -147,77 +132,39 @@
 	"bs": "波斯尼亞語",
 	"ca": "加泰羅尼亞語",
 	"ceb": "宿霧語",
-	"co": "科西嘉",
 	"cs": "捷克語",
 	"cy": "威爾士語",
 	"da": "丹麥語",
 	"de": "德語",
 	"el": "希臘語",
 	"en": "英語",
-	"en-AU": "英語（澳大利亞）",
-	"en-CA": "英語（加拿大）",
-	"en-IE": "英文（愛爾蘭）",
-	"en-IN": "英語（印度）",
-	"en-NZ": "英語（新西蘭）",
-	"en-PH": "英語（菲律賓）",
-	"en-UK": "英語（英國）",
-	"en-US": "英語（美國）",
-	"en-ZA": "英語（南非）",
 	"eo": "世界語",
 	"es": "西班牙語",
-	"es-AR": "西班牙語（阿根廷）",
-	"es-BO": "西班牙語（玻利維亞）",
-	"es-CL": "西班牙語（智利）",
-	"es-CO": "西班牙語（哥倫比亞）",
-	"es-CR": "西班牙語（哥斯達黎加）",
-	"es-DO": "西班牙語（多米尼加共和國）",
-	"es-EC": "西班牙語（厄瓜多爾）",
-	"es-ES": "西班牙語（西班牙）",
-	"es-GT": "西班牙語（危地馬拉）",
-	"es-HN": "西班牙語（洪都拉斯）",
-	"es-MX": "西班牙語（墨西哥）",
-	"es-NI": "西班牙語（尼加拉瓜）",
-	"es-PA": "西班牙語（巴拿馬）",
-	"es-PE": "西班牙語（秘魯）",
-	"es-PR": "西班牙語（波多黎各）",
-	"es-PY": "西班牙語（巴拉圭）",
-	"es-SV": "西班牙語（薩爾瓦多）",
-	"es-US": "西班牙語（美國）",
-	"es-UY": "西班牙語（烏拉圭）",
-	"es-VE": "西班牙語（委內瑞拉）",
 	"et": "愛沙尼亞語",
 	"eu": "巴斯克語",
 	"fa": "波斯語",
 	"fi": "芬蘭語",
 	"fr": "法語",
-	"fy": "弗里斯蘭",
 	"ga": "愛爾蘭語",
 	"gd": "蘇格蘭蓋爾語",
 	"gl": "加利西亞語",
 	"gu": "古吉拉特語",
-	"ha": "豪薩語",
-	"haw": "夏威夷",
 	"he": "希伯來語",
 	"hi": "印地語",
-	"hmn": "苗族",
 	"hr": "克羅地亞語",
 	"ht": "海地克里奧爾語",
 	"hu": "匈牙利語",
 	"hy": "亞美尼亞語",
 	"id": "印度尼西亞語",
-	"ig": "伊博",
 	"is": "冰島語",
 	"it": "意大利語",
-	"iw": "希伯來語",
 	"ja": "日語",
 	"jv": "爪哇語",
-	"jw": "爪哇",
 	"ka": "格魯吉亞語",
 	"kk": "哈薩克語",
 	"km": "高棉語",
 	"kn": "坎那達語",
 	"ko": "朝鮮語",
-	"ku": "庫爾德人（Kurmanji）",
 	"ky": "柯爾克孜語/吉爾吉斯語",
 	"la": "拉丁語",
 	"lb": "盧森堡語",
@@ -238,26 +185,17 @@
 	"ne": "尼泊爾語",
 	"nl": "荷蘭語",
 	"no": "挪威語",
-	"ny": "奇切瓦",
 	"pa": "旁遮普語",
 	"pap": "帕皮阿門托",
 	"pl": "波蘭語",
-	"ps": "普什圖語",
 	"pt": "葡萄牙語",
-	"pt-BR": "葡萄牙語（巴西）",
-	"pt-PT": "葡萄牙語（葡萄牙）",
 	"ro": "羅馬尼亞語",
 	"ru": "俄語",
-	"sd": "信德",
 	"si": "僧伽羅語",
 	"sk": "斯洛伐克語",
 	"sl": "斯洛維尼亞語",
-	"sm": "薩摩亞",
-	"sn": "紹納語",
-	"so": "索馬里",
 	"sq": "阿爾巴尼亞語",
 	"sr": "塞爾維亞語",
-	"st": "塞索托語",
 	"su": "巽他語",
 	"sv": "瑞典語",
 	"sw": "斯瓦希里語",
@@ -275,10 +213,7 @@
 	"vi": "越南語",
 	"xh": "科薩語",
 	"yi": "意第緒語",
-	"yo": "約魯巴",
 	"zh": "中文",
-	"zh-CN": "中國（簡體，中國）",
-	"zh-HK": "中國（繁體，香港）",
-	"zh-TW": "中國（繁體，臺灣）",
-	"zu": "祖魯"
+	"zh-CN": "中文(簡體)",
+	"zh-TW": "中文(繁體)"
 }
