+++ public/libs/preferences.js
@@ -66,7 +66,6 @@ const defaultPreferences = {
   theme: process.platform === 'darwin' ? 'systemDefault' : 'light',
   translateClipboardOnShortcut: false,
   translateWhenPressingEnter: true,
-  preferredTranslationService: 'google',
 };
 
 const getPreferences = () => ({ ...defaultPreferences, ...settings.get(`preferences.${v}`) });
