+++ src/helpers/language-utils.js
@@ -1,24 +1,5 @@
 const data = {
-  google: [
-    'auto', 'af', 'sq', 'am', 'ar', 'ar-DZ', 'ar-BH', 'ar-EG', 'ar-IL',
-    'ar-JO', 'ar-KW', 'ar-LB', 'ar-MA', 'ar-OM', 'ar-PS', 'ar-QA', 'ar-SA',
-    'ar-TN', 'ar-AE', 'hy', 'az', 'eu', 'be', 'bn', 'bs', 'bg', 'ca',
-    'ceb', 'ny', 'zh', 'zh-CN', 'zh-HK', 'zh-TW', 'co', 'hr',
-    'cs', 'da', 'nl', 'en', 'en-AU', 'en-CA', 'en-IN', 'en-IE', 'en-NZ',
-    'en-PH', 'en-ZA', 'en-UK', 'en-US', 'eo', 'et', 'tl', 'fi', 'fr',
-    'fy', 'gl', 'ka', 'de', 'el', 'gu', 'ht', 'ha', 'haw', 'iw',
-    'hi', 'hmn', 'hu', 'is', 'ig', 'id', 'ga', 'it', 'ja', 'jw',
-    'kn', 'kk', 'km', 'ko', 'ku', 'ky', 'lo', 'la', 'lv', 'lt',
-    'lb', 'mk', 'mg', 'ms', 'ml', 'mt', 'mi', 'mr', 'mn', 'my',
-    'ne', 'no', 'ps', 'fa', 'pl', 'pt', 'pt-BR', 'pt-PT', 'pa',
-    'ro', 'ru', 'sm', 'gd', 'sr', 'st', 'sn', 'sd', 'si', 'sk',
-    'sl', 'so', 'es', 'es-AR', 'es-BO', 'es-CL', 'es-CO', 'es-CR',
-    'es-DO', 'es-EC', 'es-SV', 'es-GT', 'es-HN', 'es-MX', 'es-NI',
-    'es-PA', 'es-PY', 'es-PE', 'es-PR', 'es-ES', 'es-US', 'es-UY',
-    'es-VE', 'su', 'sw', 'sv', 'tg', 'ta', 'te', 'th', 'tr', 'uk',
-    'ur', 'uz', 'vi', 'cy', 'xh', 'yi', 'yo', 'zu',
-  ],
-  yandex: [
+  all: [
     'auto',
     'af', 'am', 'ar', 'az', 'ba', 'be', 'bg', 'bn', 'bs',
     'ca', 'ceb', 'cs', 'cy', 'da', 'de', 'el', 'en', 'eo',
@@ -110,7 +91,6 @@ const data = {
     zh: 'chs',
   },
 };
-data.all = Array.from(new Set(data.google.concat(data.yandex)));
 
 export const toCountryFreeLanguage = (lang) => {
   const i = lang.indexOf('-');
@@ -130,12 +110,10 @@ export const isInput = (lang) => !(data.all.indexOf(lang) > -1);
 export const isOutput = (lang) => lang !== 'auto';
 
 // Check if language is supported by provider
-export const isGoogleSupported = (lang) => (data.google.indexOf(lang) > -1);
 export const isYandexSupported = (lang) => (data.yandex.indexOf(lang) > -1);
 
 // Get list of all languages
 export const getLanguages = () => data.all;
-export const getGoogleLanguages = () => data.google;
 export const getYandexLanguages = () => data.yandex;
 
 // Get list of all output languages
