+++ src/senders/index.js
@@ -14,7 +14,3 @@ export const requestShowRequireRestartDialog = () => ipcRenderer.send('request-s
 // Locale
 export const getLocale = (id) => ipcRenderer.sendSync('get-locale', id);
 export const getLocales = () => ipcRenderer.sendSync('get-locales');
-
-// Translate
-export const translateWithGoogleAsync = (...args) => ipcRenderer.invoke('translate-with-google-async', ...args);
-export const ttsWithGoogleAsync = (...args) => ipcRenderer.invoke('tts-with-google-async', ...args);
