--- src/setupProxy.js
@@ -1,10 +1,0 @@
-const proxy = require('http-proxy-middleware');
-
-module.exports = (app) => {
-  app.use(proxy('/translate_tts',
-    {
-      target: 'https://translate.google.com',
-      changeOrigin: true,
-      ws: true,
-    }));
-};
